📝 :: SignUpViewModel
